+++ packages/action-bar/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
 
+## 1.4.0
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @spectrum-web-components/action-group@1.4.0
+    -   @spectrum-web-components/popover@1.4.0
+    -   @spectrum-web-components/button@1.4.0
+    -   @spectrum-web-components/field-label@1.4.0
+    -   @spectrum-web-components/base@1.4.0
+
 ## 1.3.0
 
 ### Patch Changes